Move AstEq and AST visitors to their own modules.
